Bug Bytes/requirements-dev.txt
black
isort
ruff
mypy
django-stubs
mypy-extensions
djangorestframework-stubs




django
djangorestframework
django_extensions
Pillow
psycopg2-binary
